Add render optimizations and memory management improvements (#150)
* Enhanced AvaloniaBitmapTwoLayerRenderTarget and SkiaCanvasTwoLayerRenderTarget for improved pixel processing efficiency.

* Updated Vic2Rasterizer to utilize uint arrays for direct pixel manipulation, reducing casting overhead.

* Implemented zero-copy memory management with NonOwningMemoryOwner to optimize buffer handling in rendering.

+++ src/libraries/Highbyte.DotNet6502.Systems/Rendering/VideoFrameProvider/IVideoFrameLayerProvider.cs
@@ -18,5 +18,6 @@ public interface IVideoFrameLayerProvider : IVideoFrameProvider
     IReadOnlyList<LayerInfo> Layers { get; }
 
     /// Current front buffers for each layer (index matches Layers).
-    IReadOnlyList<ReadOnlyMemory<byte>> CurrentFrontLayerBuffers { get; }
+    /// Uses uint[] for direct 32-bit pixel access without casting overhead.
+    IReadOnlyList<ReadOnlyMemory<uint>> CurrentFrontLayerBuffers { get; }
 }
